update with modification for avatar upload

+++ app/Http/Controllers/Ajax/AccountController.php
@@ -25,16 +25,16 @@ class AccountController extends Controller
         if($isValid){
             $extension = $request->file('avatar')->extension();
         }else{
-            return ResponseModel::err(403,'INVALID FILE');
+            return ResponseModel::err(1004);
         }
 
-        $allow_extension = ['jpg','png','jpeg','gif'];
+        $allow_extension = ['jpg','png','jpeg','gif','bmp'];
         if($isValid && in_array($extension,$allow_extension)){
             $path = $request->file('avatar')->store('/static/img/avatar','NOJPublic');
 
             $user = Auth::user();
             $old_path = $user->avatar;
-            if($old_path != '/static/img/avatar/default.png'){
+            if($old_path != '/static/img/avatar/default.png' && $old_path != '/static/img/avatar/noj.png'){
                 Storage::disk('NOJPublic')->delete($old_path);
             }
 
@@ -43,7 +43,7 @@ class AccountController extends Controller
 
             return ResponseModel::success(200, null, '/'.$path);
         }else{
-            return ResponseModel::err(403,'INVALID FILE');
+            return ResponseModel::err(1004);
         }
 
     }
